refactor: change .js file to .ts
